fix: включение шедулеров раздельно в application.yaml
