+++ src/main/java/com/itmentorcommunityplatform/job_market_analytics_service/JobMarketAnalyticsServiceApplication.java
@@ -2,8 +2,10 @@ package com.itmentorcommunityplatform.job_market_analytics_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class JobMarketAnalyticsServiceApplication {
 
 	public static void main(String[] args) {
