+++ src/main/java/com/itmentorcommunityplatform/job_market_analytics_service/config/SchedulerConfig.java
@@ -4,8 +4,9 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@Configuration
-@EnableScheduling
-@ConditionalOnProperty(name = "scheduler.enabled", matchIfMissing = true)
+//@Configuration
+//@EnableScheduling
+//@ConditionalOnProperty(name = "scheduler.enabled", matchIfMissing = true)
 public class SchedulerConfig {
+    //TODO удалить
 }
